Fix overflow menu
