+++ canvas_modules/harness/src/client/components/custom-canvases/streams/streams-canvas.jsx
@@ -18,6 +18,7 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import { CommonCanvas, CanvasController } from "common-canvas"; // eslint-disable-line import/no-unresolved
+import { CharacterSentenceCase32, CharacterWholeNumber32 } from "@carbon/icons-react";
 
 import StreamsCanvasFlow from "./streams.json";
 
@@ -39,6 +40,7 @@ export default class StreamsCanvas extends React.Component {
 			enableAutoLayoutHorizontalSpacing: 80,
 			enableInternalObjectModel: true,
 			enableMoveNodesOnSupernodeResize: true,
+			enableContextToolbar: true,
 			enableDropZoneOnExternalDrag: true,
 			enableNarrowPalette: false,
 			schemaValidation: true,
@@ -72,12 +74,44 @@ export default class StreamsCanvas extends React.Component {
 		});
 	}
 
+	contextMenuHandler(source, defaultMenu) {
+		const defMenu = defaultMenu;
+		if (source.type === "comment") {
+			const lettersSubMenu = [
+				{ action: "A", label: "Aaaaaaaa" },
+				{ action: "B", label: "Bbbb" },
+				{ action: "C", label: "Cccccccccc" },
+				{ action: "D", label: "Ddd" },
+				{ action: "E", label: "Eeeeee" },
+				{ action: "F", label: "Ffffffff" },
+				{ action: "G", label: "Gg" },
+				{ action: "H", label: "Hhhhh" },
+				{ action: "I", label: "Iiiiiii" }
+			];
+			const numbersSubMenu = [
+				{ action: "1", label: "One 1" },
+				{ action: "2", label: "Two Two" },
+				{ action: "3", label: "Three Three" },
+				{ action: "4", label: "Four" },
+				{ action: "5", label: "Five 555" },
+				{ action: "6", label: "Sixes" },
+				{ action: "7", label: "Seven" },
+				{ action: "8", label: "Eight" },
+				{ action: "9", label: "Niners 9" }
+			];
+			defMenu.push({ action: "letters", label: "Letters", submenu: true, menu: lettersSubMenu, icon: (<CharacterSentenceCase32 />), toolbarItem: true });
+			defMenu.push({ action: "numbers", label: "Numbers", submenu: true, menu: numbersSubMenu, icon: (<CharacterWholeNumber32 />) });
+		}
+		return defMenu;
+	}
+
 	render() {
 
 		return (
 			<CommonCanvas
 				canvasController={this.canvasController}
 				config={this.config}
+				contextMenuHandler={this.contextMenuHandler}
 			/>
 		);
 	}
